pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Oleg <chaiko.oleg89@gmail.com>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.2"
flake8 = "^7.0.0"
pytest = "^8.0.1"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
validators = "^0.22.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pytest = "^8.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
